requirements.txt
matplotlib==3.5.1
numpy==1.21.1
scipy==1.7.3
spatialmath-python==0.11
sympy==1.9
